+++ m-common/src/main/java/com/antmen/antwork/common/api/controller/ReservationController.java
@@ -1,0 +1,60 @@
+package com.antmen.antwork.common.api.controller;
+
+import com.antmen.antwork.common.api.request.ReservationRequestDto;
+import com.antmen.antwork.common.api.request.ReservationStatusChangeRequestDto;
+import com.antmen.antwork.common.api.request.ReservationCancelRequestDto;
+import com.antmen.antwork.common.api.response.ReservationResponseDto;
+import com.antmen.antwork.common.service.ReservationService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/reservation")
+public class ReservationController {
+
+    private final ReservationService reservationService;
+
+    /**
+     * 예약 생성
+     */
+    @PostMapping
+    public ResponseEntity<ReservationResponseDto> createReservation(
+            @RequestBody ReservationRequestDto requestDto) {
+        ReservationResponseDto responseDto = reservationService.createReservation(requestDto);
+        return ResponseEntity.ok(responseDto);
+    }
+
+    /**
+     * 예약 단건 조회
+     */
+    @GetMapping("/{id}")
+    public ResponseEntity<ReservationResponseDto> getReservation(
+            @PathVariable Long id) {
+        ReservationResponseDto responseDto = reservationService.getReservation(id);
+        return ResponseEntity.ok(responseDto);
+    }
+
+    /**
+     * 예약 상태 변경
+     */
+    @PatchMapping("/{id}/status")
+    public ResponseEntity<Void> changeStatus(
+            @PathVariable Long id,
+            @RequestBody ReservationStatusChangeRequestDto dto) {
+        reservationService.changeStatus(id, dto.getStatus());
+        return ResponseEntity.ok().build();
+    }
+
+    /**
+     * 예약 취소
+     */
+    @PostMapping("/{id}/cancel")
+    public ResponseEntity<Void> cancelReservation(
+            @PathVariable Long id,
+            @RequestBody ReservationCancelRequestDto dto) {
+        reservationService.cancelReservation(id, dto.getCancelReason());
+        return ResponseEntity.ok().build();
+    }
+}
