+++ m-common/src/main/java/com/antmen/antwork/common/service/ReservationService.java
@@ -30,36 +30,27 @@ public class ReservationService {
             ReservationConstants.STATUS_ERROR
     );
 
-    @Transactional
+    /**
+     * 예약 생성
+     */
     public ReservationResponseDto createReservation(ReservationRequestDto requestDto) {
-        try {
             Reservation reservation = reservationMapper.toEntity(requestDto);
             Reservation saved = reservationRepository.save(reservation);
             return reservationMapper.toDto(saved);
-        } catch (Exception e) {
-            throw new RuntimeException("예약 생성 중 오류가 발생했습니다.", e);
-        }
     }
-
-
-    public void changeReservationStatus__(Long id, String newStatus) {
-        Reservation reservation = reservationRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("해당 예약이 존재하지 않습니다."));
-        reservation.setReservationStatus(newStatus);
-    }
-
 
     /**
-     * 예약 취소 처리
+     * 예약 단건 조회
      */
-    public void cancelReservation__(Long id, String cancelReason) {
+    public ReservationResponseDto getReservation(Long id) {
         Reservation reservation = reservationRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("해당 예약이 존재하지 않습니다."));
-        reservation.setReservationStatus("C");
-        reservation.setReservationCancelReason(cancelReason);
+                return reservationMapper.toDto(reservation);
     }
 
-    @Transactional
+    /**
+     * 예약 상태 변경
+     */
     public void changeStatus(Long id, String status) {
         if (!VALID_STATUS.contains(status)) {
             throw new IllegalArgumentException("유효하지 않은 예약 상태입니다.");
@@ -70,12 +61,21 @@ public class ReservationService {
         reservationRepository.save(reservation);
     }
 
-    @Transactional
+    /**
+     * 예약 취소
+     */
     public void cancelReservation(Long id, String cancelReason) {
+
+        if (cancelReason == null || cancelReason.trim().isEmpty()) {
+            throw new IllegalArgumentException("취소 사유는 필수입니다.");
+        }
+
         Reservation reservation = reservationRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("예약을 찾을 수 없습니다."));
         reservation.setReservationStatus(ReservationConstants.STATUS_CANCEL);
         reservation.setReservationCancelReason(cancelReason);
         reservationRepository.save(reservation);
     }
+
+
 }
